+++ airbyte-integrations/connectors/destination-gcs-data-lake/src/main/kotlin/io/airbyte/integrations/destination/gcs_data_lake/catalog/GcsDataLakeNameGeneratorFactory.kt
@@ -6,21 +6,44 @@ package io.airbyte.integrations.destination.gcs_data_lake.catalog
 
 import io.airbyte.cdk.load.command.DestinationStream
 import io.airbyte.cdk.load.data.Transformations
+import io.airbyte.cdk.load.orchestration.db.ColumnNameGenerator
 import io.airbyte.cdk.load.orchestration.db.FinalTableNameGenerator
 import io.airbyte.cdk.load.table.TableName
 import io.airbyte.cdk.load.toolkits.iceberg.parquet.TableIdGenerator
 import io.airbyte.cdk.load.toolkits.iceberg.parquet.tableIdOf
 import io.airbyte.integrations.destination.gcs_data_lake.spec.GcsDataLakeConfiguration
 import io.micronaut.context.annotation.Factory
-import javax.inject.Singleton
+import jakarta.inject.Singleton
 import org.apache.iceberg.catalog.TableIdentifier
+
+/**
+ * Generates BigLake-compatible column names for GCS Data Lake tables.
+ *
+ * BigLake external tables have strict naming requirements:
+ * - Only alphanumeric characters (a-z, A-Z, 0-9) and underscores (_)
+ * - Prefixes with underscore if it starts with a number
+ *
+ * This implementation uses [Transformations.toAlphanumericAndUnderscore] to ensure all column names
+ * meet these requirements.
+ */
+@Singleton
+class GcsDataLakeColumnNameGenerator : ColumnNameGenerator {
+    override fun getColumnName(column: String): ColumnNameGenerator.ColumnName {
+        val sanitized = sanitizeBigLakeName(column)
+        return ColumnNameGenerator.ColumnName(
+            displayName = sanitized,
+            canonicalName = sanitized,
+        )
+    }
+}
+
 
 /**
  * Sanitizes a name to be BigLake/BigQuery compatible.
  * - Converts to alphanumeric + underscore only
  * - Prefixes with underscore if it starts with a number
  */
-private fun sanitizeBigLakeTableName(name: String): String {
+private fun sanitizeBigLakeName(name: String): String {
     var sanitized = Transformations.toAlphanumericAndUnderscore(name)
     // BigLake/BigQuery doesn't allow table names starting with numbers
     // Prefix with underscore if it starts with a digit
@@ -37,14 +60,14 @@ private fun sanitizeBigLakeTableName(name: String): String {
  */
 class BigLakeTableIdGenerator(private val databaseName: String) : TableIdGenerator {
     override fun toTableIdentifier(stream: DestinationStream.Descriptor): TableIdentifier {
-        val namespace = sanitizeBigLakeTableName(stream.namespace ?: databaseName)
-        val name = sanitizeBigLakeTableName(stream.name)
+        val namespace = sanitizeBigLakeName(stream.namespace ?: databaseName)
+        val name = sanitizeBigLakeName(stream.name)
         return tableIdOf(namespace, name)
     }
 }
 
 @Factory
-class GcsDataLakeTableIdGeneratorFactory(
+class GcsDataLakeNameGeneratorFactory(
     private val gcsDataLakeConfiguration: GcsDataLakeConfiguration
 ) {
     @Singleton
@@ -59,8 +82,8 @@ class GcsDataLakeTableIdGeneratorFactory(
     fun createFinalTableNameGenerator(): FinalTableNameGenerator =
         FinalTableNameGenerator { stream ->
             val namespace =
-                sanitizeBigLakeTableName(stream.namespace ?: gcsDataLakeConfiguration.namespace)
-            val name = sanitizeBigLakeTableName(stream.name)
+                sanitizeBigLakeName(stream.namespace ?: gcsDataLakeConfiguration.namespace)
+            val name = sanitizeBigLakeName(stream.name)
             TableName(namespace = namespace, name = name)
         }
 }
